fixed some small errors and made it so the made the game more readable and playable
fixed an error when the client lost from quessing too many times the categories they beat would be wouldn't be reset for the next game. made it so server tells client how many letters are in the word. tell client if they won, lost, guess word right or wrong or ran out of guesses.